--- a/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/IHubProtocol.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/IHubProtocol.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.Buffers;
 using Microsoft.AspNetCore.Connections;
 
@@ -18,7 +19,7 @@ namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
 
         void WriteMessage(HubMessage message, IBufferWriter<byte> output);
 
-        byte[] GetMessageBytes(HubMessage message);
+        ReadOnlyMemory<byte> GetMessageBytes(HubMessage message);
 
         bool IsVersionSupported(int version);
     }