--- a/src/Microsoft.AspNetCore.SignalR.Core/Internal/SerializedMessage.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Core/Internal/SerializedMessage.cs
@@ -1,11 +1,13 @@
+using System;
+
 namespace Microsoft.AspNetCore.SignalR.Internal
 {
     public readonly struct SerializedMessage
     {
         public string ProtocolName { get; }
-        public byte[] Serialized { get; }
+        public ReadOnlyMemory<byte> Serialized { get; }
 
-        public SerializedMessage(string protocolName, byte[] serialized)
+        public SerializedMessage(string protocolName, ReadOnlyMemory<byte> serialized)
         {
             ProtocolName = protocolName;
             Serialized = serialized;