--- a/src/Microsoft.AspNetCore.SignalR.Core/Internal/SerializedMessage.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Core/Internal/SerializedMessage.cs
@@ -1,3 +1,6 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 
 namespace Microsoft.AspNetCore.SignalR.Internal