- Changed GetMessageBytes to return ReadOnlyMemory
- Make HandshakeProtocol.SuccessHandshakeData a readonly field

## Before

```
             Method |       Input | HubProtocol |       Mean |     Error |    StdDev |        Op/s |  Gen 0 | Allocated |
------------------- |------------ |------------ |-----------:|----------:|----------:|------------:|-------:|----------:|
 WriteSingleMessage | NoArguments |        Json | 1,074.8 ns | 11.669 ns | 10.915 ns |   930,399.5 | 0.0114 |     392 B |
 WriteSingleMessage | NoArguments |     MsgPack |   344.3 ns |  7.487 ns |  6.637 ns | 2,904,147.3 | 0.0010 |      40 B |
```

## After

```
              Method |       Input | HubProtocol |       Mean |     Error |    StdDev |        Op/s |  Gen 0 | Allocated |
------------------- |------------ |------------ |-----------:|----------:|----------:|------------:|-------:|----------:|
 WriteSingleMessage | NoArguments |        Json | 1,112.9 ns | 12.325 ns | 11.529 ns |   898,513.8 | 0.0114 |     392 B |
 WriteSingleMessage | NoArguments |     MsgPack |   366.7 ns |  7.010 ns |  7.791 ns | 2,726,694.7 | 0.0010 |      40 B |
```

There's a performance hit here, but it might still be worth it?