@BrennanConroy 2018-04-13
Why did it get slower?