@davidfowl 2018-04-13
We're creating a `ROMemory<byte>` over the array every time, that's not free.